pyproject.toml
[tool.poetry]
name = "social-app"
version = "0.1.0"
description = ""
authors = ["otabek <gulomovotabek444@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
pillow = "^11.0.0"
django-debug-toolbar = "^4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
